fix: Rename analysis-utils.ts to analysis_utils.ts for Convex compatibility
- Convex module filenames can only contain alphanumeric characters, underscores, or periods
- Updated all import statements to reference the new filename
- Fixed BadConvexModuleIdentifier error when running convex dev
