+++ .taskmaster/tasks/tasks.json
@@ -335,7 +335,7 @@
             "description": "Connect panel to live analyses data using Convex queries with real-time updates",
             "dependencies": ["5.1"],
             "details": "Create useQuery hook for analyses table subscription, implement real-time updates when new analyses arrive, add loading states during analysis processing, ensure data fetching linked to current messageId",
-            "status": "pending",
+            "status": "in-progress",
             "testStrategy": ""
           },
           {
@@ -709,7 +709,7 @@
     ],
     "metadata": {
       "created": "2025-07-23T19:21:41.002Z",
-      "updated": "2025-07-24T16:34:49.577Z",
+      "updated": "2025-07-24T16:59:45.031Z",
       "description": "Tasks for master context"
     }
   }
